+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Updated dependencies
+
+### Fixed
+- Changed ip in tests from `0.0.0.0` to `localhost` to fix them
+  running on Windows.
+
 ## [1.4.0] - 2019-07-13
 ### Added
 - `json-using-serde` feature.